Seccion Categoria

+++ Tasker/MVVM/Models/Category.cs
@@ -16,7 +16,7 @@ namespace Tasker.MVVM.Models
 
         public int PendingTasks { get; set; }
 
-        public int Percentage { get; set; }
+        public float Percentage { get; set; }
 
     }
 }
